+++ scelvis/store.py
@@ -5,13 +5,24 @@ and memoizes the loaded data in the Flask cacche.  This module is aware of the F
 thus also of the Dash app.
 """
 
-import glob
 import os.path
+import urllib
 
+import fs
 from logzero import logger
 
 from . import data, settings
 from .cache import cache
+
+
+def get_file_system(url):
+    """Return file system from the given ``url``."""
+    url = urllib.parse.urlparse(url)
+    if url.scheme and url.scheme != "file":
+        base_url = "%s://%s" % (url.scheme, url.netloc)
+    else:
+        base_url = url.netloc
+    return url, fs.open_fs(base_url)
 
 
 @cache.memoize()
@@ -21,38 +32,42 @@ def load_all_metadata():
     If ``data_dir`` itself contains a file called ``about.md``, it is assumed that only one dataset is available.
     Otherwise, all sub directories will be scanned for ``about.md`` and one entry is returned for each dataset.
     """
-    data_dir = settings.DATA_DIR
-    if os.path.exists(os.path.join(data_dir, settings.ABOUT_FILENAME)):
-        logger.info("Loading single dataset from data directory %s", data_dir)
-        return [data.load_metadata(data_dir)]
-    else:
-        result = []
-        for path in glob.glob(os.path.join(data_dir, "*", settings.ABOUT_FILENAME)):
-            result.append(data.load_metadata(os.path.dirname(path)))
-        if result:
-            logger.info("Loaded %d data sets from data directory.", len(result))
+    result = []
+    for data_dir in settings.DATA_DIRS:
+        url, the_fs = get_file_system(data_dir)
+        if the_fs.exists(fs.path.join(url.path, settings.ABOUT_FILENAME)):
+            logger.info("Loading single dataset from data directory %s", data_dir)
+            result.append(data.load_metadata(data_dir))
         else:
-            logger.warn("No data sets found in data directory %s", data_dir)
-        return result
+            for match in the_fs.glob(fs.path.join(url.path, "*", settings.ABOUT_FILENAME)):
+                logger.info("Found data set %s", fs.path.dirname(match.path))
+                result.append(data.load_metadata(the_fs, fs.path.dirname(match.path)))
+            if result:
+                logger.info("Loaded %d data sets from data directory.", len(result))
+            else:
+                logger.warn("No data sets found in data directory %s", data_dir)
+    return result
 
 
 @cache.memoize()
 def load_metadata(identifier):
     """Load metadata for the given identifier from data or upload directory."""
-    for base_dir in (settings.DATA_DIR, settings.UPLOAD_DIR):
+    for base_dir in settings.DATA_DIRS + [settings.UPLOAD_DIR]:
         if not base_dir:
             continue
-        full_path = os.path.join(base_dir, identifier)
-        if os.path.exists(full_path):
-            return data.load_metadata(full_path)
+        url, the_fs = get_file_system(base_dir)
+        full_path = fs.path.join(url.path, identifier)
+        if the_fs.exists(full_path):
+            return data.load_metadata(the_fs, full_path)
 
 
 @cache.memoize()
 def load_data(identifier):
     """Load data for the given identifier from data or upload directory."""
-    for base_dir in (settings.DATA_DIR, settings.UPLOAD_DIR):
+    for base_dir in settings.DATA_DIRS + [settings.UPLOAD_DIR]:
         if not base_dir:
             continue
-        full_path = os.path.join(base_dir, identifier)
-        if os.path.exists(full_path):
-            return data.load_data(full_path)
+        url, the_fs = get_file_system(base_dir)
+        full_path = fs.path.join(url.path, identifier)
+        if the_fs.exists(full_path):
+            return data.load_data(the_fs, full_path)