+++ HISTORY.rst
@@ -10,6 +10,7 @@ v0.2.0 (unreleased)
 - Fixing dependency on ``ruamel-yaml`` to ``ruamel.yaml``.
 - Adding conversion feature.
 - Adding upload feature.
+- Adding support from SSHFS and FTP through pyfilesystem (no FTPS support).
 
 ------
 v0.1.0